Fixed a conversion warning.
A "possible loss of data" warning was triggered with MSVC since
`process.run(int)` implicitly converted an unsigned `size_t` to a signed
`int`. This patch fixes the type of the passed values.

Review: https://reviews.apache.org/r/65326

+++ 3rdparty/libprocess/src/tests/benchmarks.cpp
@@ -674,10 +674,10 @@ private:
 // Measures performance of message passing in ProtobufProcess.
 TEST(ProcessTest, Process_BENCHMARK_ProtobufInstallHandler)
 {
-  const size_t submessages[] = {0, 1, 5, 10, 50, 100, 500, 1000, 5000, 10000};
+  const int submessages[] = {0, 1, 5, 10, 50, 100, 500, 1000, 5000, 10000};
 
   ProtobufInstallHandlerBenchmarkProcess process;
-  foreach (size_t num_submessages, submessages) {
+  foreach (int num_submessages, submessages) {
     process.run(num_submessages);
   }
 }
